qb-io async/transport: protocol-switch UAF guard, DoS and fail-closed hardening
- io.h / buffered_io.h (process_messages / process_input): snapshot the active
  protocol's should_flush() into a local BEFORE onMessage() instead of reading
  it back off the saved (old) protocol pointer afterwards. onMessage() may
  switch_protocol() or clear_protocols(), either of which can leave that pointer
  dangling; should_flush() is a per-protocol invariant (no setter is ever
  called), so the snapshot is equivalent and removes every post-onMessage deref.

- protocol/json.h: msgpack (json_packed) had no nesting-depth guard before
  from_msgpack() — a recursive reader stack-overflow DoS. Add a SAX depth pre-pass
  (the sibling text json already had json_depth_within).

- transport/accept.h + saccept.h: the acceptor was disposed permanently on
  EMFILE/ENFILE/ENOBUFS/ENOMEM (only ECONNABORTED/EPROTO were retried) — an
  fd-exhaustion DoS. Remap those transient errnos to EWOULDBLOCK.

- transport/udp.h: ProxyOut wild-pointer fix on the header-allocate path,
  0-length datagram setDestination, write() doc.

- system/sys__ifaddrs.h: fill_sa_address default branch had an assert-only bound
  (heap overflow in release) — add a runtime check mirroring fill_ll_address.

- tcp/ssl/socket.cpp: apply_client_peer_verification left X509_VERIFY_PARAM_set1_*
  return values unchecked (fail-OPEN hostname verification); fail closed. SSL
  write ret<=0 handling.

- sys__socket.cpp (strerror thread-safety), compression.cpp (size==0 infinite
  loop guard), nanolog.cpp (sign-compare), io_handler.h (broadcast re-entrancy /
  DoS cap), config.h (doc). test-io.cpp: MsgpackDepthGuard coverage.

+++ modules/nanolog/nanolog.cpp
@@ -620,7 +620,10 @@ public:
         auto pos = m_os->tellp();
         logline.stringify(*m_os);
         m_bytes_written += m_os->tellp() - pos;
-        if (m_bytes_written > m_log_file_roll_size_bytes) {
+        // m_bytes_written (std::streamoff, signed) is always >= 0 here; cast to
+        // match the unsigned roll-size and avoid a signed/unsigned comparison
+        // warning (the roll size became uint64_t to fix the >=4GB overflow).
+        if (static_cast<std::uint64_t>(m_bytes_written) > m_log_file_roll_size_bytes) {
             roll_file();
         }
     }
